Moves code into domain-model
Mostly trying to figure out the right spot to put all this data.
