+++ src/extension/domains/campaign.ts
@@ -1,7 +1,7 @@
-import { replicate } from "./utils";
+import { replicate } from "../utils";
 
 import type { NodeCG, Replicant } from "nodecg-types/types/server";
-import type { CampaignClient } from "./clients/tiltify-client";
+import type { CampaignClient } from "../clients/tiltify-client";
 
 export default (nodecg: NodeCG, client: CampaignClient) => {
   nodecg.log.info("⬆ Starting Campaign client...");
