+++ src/extension/domains/milestones.ts
@@ -1,16 +1,13 @@
-import { replicate } from "./utils";
+import { replicate } from "../utils";
 
 import type { NodeCG, Replicant } from "nodecg-types/types/server";
-import type { CampaignClient } from "./clients/tiltify-client";
+import type { CampaignClient } from "../clients/tiltify-client";
 
 export default (nodecg: NodeCG, client: CampaignClient) => {
   nodecg.log.info("⬆ Starting Milestone client...");
-  const milestones: Replicant<Array<object>> = nodecg.Replicant(
-    "milestones",
-    {
-      defaultValue: [],
-    }
-  );
+  const milestones: Replicant<Array<object>> = nodecg.Replicant("milestones", {
+    defaultValue: [],
+  });
 
   client.on("getMilestones", replicate(milestones));
 };
